requirements.txt
textual>=0.40.0
rich>=13.0.0
fabric>=3.2.0
paramiko>=3.4.0
pyyaml>=6.0
pydantic>=2.0
pydantic-settings>=2.0
cryptography>=41.0.0
requests>=2.31.0
